Address comments, move to own class due to circular import issues
